eval/requirements.txt
playwright>=1.40.0
anthropic>=0.39.0
requests>=2.31.0
Pillow>=10.0.0
